requirements-test.txt
pytest>=7.0.0
pyyaml>=6.0
jsonschema>=4.0.0
pydantic>=1.10.0
pre-commit>=3.0.0
